Move recommendation functions to separate file
