+++ project/frontend/src/FindMatch/PreferenceForm.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import getRecommendation from './Recommender.js';
 
 class PreferenceForm extends React.Component {
   constructor(props) {
@@ -41,16 +42,15 @@ class PreferenceForm extends React.Component {
       this.setState({ [event.target.name]: field });
     } else if (event.target.name === 'diningExp') {
       field['pref'] = event.target.value;
-      this.setState({ [event.target.name]: field })
+      this.setState({ [event.target.name]: field });
     } else {
       field['pref'] = parseInt(event.target.value);
-      this.setState({ [event.target.name]: field })
+      this.setState({ [event.target.name]: field });
     }
-
   }
 
   handleSubmit(event) {
-    this.getRecommendation();
+    getRecommendation(/* preferences= */ this.state);
     event.preventDefault();
   }
 
@@ -166,82 +166,6 @@ class PreferenceForm extends React.Component {
         <button type='submit'>Submit</button>
       </form>
     );
-  }
-
-  getRecommendation() {
-    const promises = this.makePromisesArray();
-
-    Promise.all(promises)
-      // This gives us the list of restaurants.
-      .then((responses) =>
-        Promise.all(responses.map((response) => response.json()))
-      )
-      .then((data) => {
-        let restaurants = [];
-        for (const restaurant of data) {
-          restaurants = restaurants.concat(restaurant.results);
-        }
-
-        fetch('/api/recommendation', {
-          method: 'POST',
-          headers: {
-            'Access-Control-Allow-Origin': '*',
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            restaurants,
-            preferences: this.state,
-          }),
-        })
-          .then((response) => response.text())
-          .then((data) => {
-            try {
-              const selections = JSON.parse(data);
-              console.log(selections);
-            } catch (err) {
-              console.log(err);
-            }
-          });
-      });
-  }
-
-  /**
-   *  Returns an array of promises of calls to the Google Places API.
-   *  One promise is created for every cuisine type.
-   *  If no cuisines are specified, only one promise is created.
-   */
-  makePromisesArray() {
-    const { cuisine, radius, currLocation, open } = this.state;
-    // TODO: replace API key
-    const apiKey = 'AIzaSyBBqtlu5Y3Og7lzC1WI9SFHZr2gJ4iDdTc';
-    const proxyUrl = 'https://cors-anywhere.herokuapp.com/';
-    const textSearchBaseUrl =
-      'https://maps.googleapis.com/maps/api/place/textsearch/json?';
-    const promises = [];
-    const cuisineTypes = cuisine;
-    // Make sure we still create a promise even if no cuisine type is specified.
-    // This will make the text search query just "restaurant" without specifying a cuisine.
-    if (cuisineTypes.length === 0) {
-      cuisineTypes.push('');
-    }
-    for (const cuisineType of cuisineTypes) {
-      const searchParams = new URLSearchParams();
-      searchParams.append('query', cuisineType + 'restaurant');
-      searchParams.append('location', currLocation.lat + ',' + currLocation.lng);
-      if (radius.pref) {
-        searchParams.append('radius', this.milesToMeters(this.state.radius.pref));
-      }
-      if (open) {
-        searchParams.append('opennow', open);
-      }
-      searchParams.append('key', apiKey);
-      promises.push(fetch(proxyUrl + textSearchBaseUrl + searchParams));
-    }
-    return promises;
-  }
-
-  milesToMeters(numMiles) {
-    return numMiles * 1609.34;
   }
 }
